backend/requirements.txt
# Django Backend - Resume Critique Agent
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.59.0
dj-rest-auth[with_social]==5.0.2
PyJWT==2.8.0

# Async Task Processing
celery==5.3.4
redis==5.0.1

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# NLP & Machine Learning
spacy==3.7.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.3

# LangChain & LangGraph for Multi-Agent System
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langgraph==0.0.20
langchain-openai==0.0.2
langchain-huggingface==0.0.1

# Testing
pytest==7.4.4
pytest-django==4.7.0

# Code Quality
flake8==7.0.0
